copies: move `CopyRecords` to new `copies` module
Copy/rename handling is complicated. It seems worth having a module
for it. I'm going to add more content to it next.
